refactor(profile/reports): move `listeningDays` to `views`
